client/ai_ops/requirements.txt
#requirements

yfinance
asyncio
langchain-mcp-adapters
langchain-openai
langgraph
markitdown>=0.0.1a3
oracledb>=2.0.0
python-dotenv>=1.0.0
streamlit
appdirs
oci
oracle-ads
langchain
langchain-core
langchain_community
mcp>=1.11,<1.12 
fastmcp>=2.5,<2.6
uvicorn>=0.30.0 
anyio>=4.4,<5 
httpx>=0.27
plotly
matplotlib
geopandas
streamlit
streamlit-autorefresh
graphviz